// src/app/core/models/pokemon.model.ts
export interface Pokemon {
    id: number;
    name: string;
    base_experience: number;
    height: number;
    is_default: boolean;
    order: number;
    weight: number;
    abilities: Ability[];
    forms: Form[];
    game_indices: Index[];
    held_items: HeldItem[];
    location_area_encounters: string;
    moves: Mfe[];
    species: Species;
    sprites: Sprites;
    cries: Cries;
    stats: Stat[];
    types: Type[];
    past_types: PastType[];
    description:'';
  }
  
  export interface Ability {
    is_hidden: boolean;
    slot: number;
    ability: Ability2;
  }
  
  export interface Ability2 {
    name: string;
    url: string;
  }
  
  export interface Form {
    name: string;
    url: string;
  }
  
  export interface Index {
    game_index: number;
    version: Version;
  }
  
  export interface Version {
    name: string;
    url: string;
  }
  
  export interface HeldItem {
    item: Item;
    version_details: VersionDetail[];
  }
  
  export interface Item {
    name: string;
    url: string;
  }
  
  export interface VersionDetail {
    rarity: number;
    version: Version2;
  }
  
  export interface Version2 {
    name: string;
    url: string;
  }
  
  export interface Mfe {
    move: Move;
    version_group_details: VersionGroupDetail[];
  }
  
  export interface Move {
    name: string;
    url: string;
  }
  
  export interface VersionGroupDetail {
    level_learned_at: number;
    version_group: VersionGroup;
    move_learn_method: MoveLearnMethod;
  }
  
  export interface VersionGroup {
    name: string;
    url: string;
  }
  
  export interface MoveLearnMethod {
    name: string;
    url: string;
  }
  
  export interface Species {
    name: string;
    url: string;
  }
  
  export interface Sprites {
    back_default: string;
    back_female: any;
    back_shiny: string;
    back_shiny_female: any;
    front_default: string;
    front_female: any;
    front_shiny: string;
    front_shiny_female: any;
    other: Other;
    versions: Versions;
  }
  
  export interface Other {
    dream_world: DreamWorld;
    home: Home;
    'official-artwork': OfficialArtwork;
    showdown: Showdown;
  }
// ...
  export interface Versions {
    'generation-i': GenerationI;
    'generation-ii': GenerationIi;
    'generation-iii': GenerationIii;
    'generation-iv': GenerationIv;
    'generation-v': GenerationV;
    'generation-vi': GenerationVi;
    'generation-vii': GenerationVii;
    'generation-viii': GenerationViii;
  }
// ...
  export interface Cries {
    latest: string;
    legacy: string;
  }
  
  export interface Stat {
    base_stat: number;
    effort: number;
    stat: Stat2;
  }
  
  export interface Stat2 {
    name: string;
    url: string;
  }
  
  export interface Type {
    slot: number;
    type: Type2;
  }
  
  export interface Type2 {
    name: string;
    url: string;
  }
  
  export interface PastType {
    generation: Generation;
    types: Type3[];
  }
  
  export interface Generation {
    name: string;
    url: string;
  }
  
  export interface Type3 {
    slot: number;
    type: Type4;
  }